Higher level method 'action' to send AMI actions
The 'send' method fires the provided callback zero to many times,
for every recieved response/event with corresponding ActionID.

A method that collects all the replies and then fires callback only
once, either wirh error or with the compiled reponse, would be much more
handy.

The 'action' method works well for actions that return a single response
or an EventList or EventList-like series of events that start with a
'will follow' responce and end with the '*Complete' event.

There are some actions that do not follow the convention, however,
namely 'Originate' (may or may not send OriginateResponse),
'SIPqualifypeer' (fires extra QualifyPeerDone event, and on older
Asterisks may not even return a response), and a few others.

So I'll either leave both low-level 'send', not handy but universal,
and higher-level 'action', or add special handling of these particular
actions to the higher-level method and deprecate the old one.